+++ springapp/src/main/java/com/example/springapp/repository/MedicalRecordRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.springapp.model.MedicalRecord;
 
-public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Integer> {
+public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Long> {
 
 }
